requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
deepeval>=0.21.0
psutil>=5.9.0
tiktoken>=0.7.0
rich>=13.0.0
pyyaml>=6.0
